fix(resolver) Resolution with prefix was incorrect in some cases
Delegate computation for indicating if a path is absolute to each
resolver.
Delegate path combination with prefix/suffix to each resolver.
Add tests to ensure that resolution works as expected.
Add tests to ensure that a prefix per template engine and per resource
resolver (classpath or file) works as expected