chore(tools) Development scripts for detecting properties in code and docs replaced by Python
Shell were having an issue with broken pipes. The cause seem to be buffering issue. None of solution found on the Internet
was working. Moreover, the code was short but difficult to maintain (in one year, I would spend too much time to get
back into the code). The scripts are now totally rewritten in Python. Scripts are now clearer and much easier to
maintain.